requirements.txt
mcp>=1.26.0
pyodbc>=5.0.0
azure-identity>=1.19.0
